persistlog: fix garbage collector, Get, and tests
This commit reverts some of the work of the previous commit. It
fixes the garbage collector to sweep up the CLTV by decrementing
the CLTV for evey block notification it receives. It also fixes
the Get method to return a CLTV value of math.MaxUint32 upon
retrieving no CLTV value. Tests were also cleaned up and fixed.