persistlog: persistent GC, cleaned tests, batched writes
This commit introduces changes to the garbage collector that cause
it to batch all writes to the backend boltdb. The garbage collector's
code is much cleaner now and all the logic is contained in the Batch
call. Additionally, persistence was added to the garbage collector
so that even on DecayedLog shutdown, the garbage collector will
still delete expired entries based on their CLTV. Tests were cleaned
up and a new test for GC persistence was added.