eob: add new eob file to house packing/unpacking EOB data within onion hops
In this commit, we add a new file eob.go which will house all of our
logic for encoding and decoding any data for EOB hops. An EOB hop is a
hop that contains extra data and may span over several hops. There are
two types of EOB hops: pivot hops, and full hops. Within a pivot hop, we
can fit up 10 bytes of data as we use 1 byte to signal the type of the
EOB data, and another byte to signal if there are more hops (high bit)
and the number of bytes consumed (lower 7 bytes). Within a full hop, we
can fit up to 32 bytes as we use all the regular forwarding fields, but
leave the HMAC in tact, and then use 1 byte of the extra padding bytes
to encode more+length in a similar manner as the pivot hop.

Currently (just a draft) there are three EOB data types. Empty, sphinx
send, and inner TLV. Empty is the default value used today and indicates
that there are no extra bytes encoded. The sphinx type is used to send
non-interactive payments to a destination node. It consumes by default 2
hops and will encode a 32 byte pre-image along with 10 bytes of extra
data which can be used as an identifier. The final TLV type is mean to
signal that an upper layer will interpret the opaque bytes as encoded
TLV values.