path: extend OnionHop with EOB field, add concept of unrolling a path
In this commit, we extend the OnionHop field to allow a caller to
specify which hops should be possibly expanded to encode extra onion
blob data. We call the process of expanding a hop to encoded extra data
"unrolling". As a result of the unrolling process, the true size of a
path may increase. For example, if we have a 2 hop route, and we want to
encode 32 bytes for the last hop, since that can't fit into a pivot hop,
we need to add an additional full hop which increases the size of the
path to 3 hops.

Along the way we add some helper functions used to determine how large
an unrolled will be EOB data wise, and if it has any data at all.