sphinx: expand NewOnionPacket to unroll EOB hops
In this commit, we update the NewOnionPacket method to detect when a hop
contains EOB data. If so, then we'll unroll the entire path into one
which encodes the EOB data within the hop data of the same, or
additional virtual hops. The unrolling process will duplicate hop public
keys accordingly such that we encrypt a hop to the same public key for
several hops to allow the target node to continue to parse and unwrap
each packet.