sphinx: update processOnionPacket to be able to recursively unwrap EOB hops
In this commit, we update the processOnionPacket method to allow it to
detect if there are any additional hops encoded to the router than
contain EOB data. If so, then it will unpack the pivot hop, and then
continue to unwrap layers from the onion until all EOB data has been
recovered. Processing is mostly the same. The major difference is that
we check the HMAC at the _final_ outer hop rather than the first in
order to properly detect the exit hop.

We've also added a series of tests to ensure that we're able to uncover
EOB data at each hop, even when an EOB hop expands into multiple actual
hops.