crypto: add ECDH interface and implementation
With this commit we create a new interface that abstracts away the
ECDH implementation. We also create one implementation that satisfies
the interface that can be used if the private key is known directly.
Later we will change the generateSharedSecret function to use this
interface instead of doing the ECDH operation directly.