sphinx: add strictAttribution flag for onion error decryption
We now add a flag to the decryption function which controls the way
we're carrying out the decryption with regards to the attribution data.
This is primarily done for graceful backwards compatibility, in order to
prevent nodes upgrading to attributable failures from blaming their
first hop peers, which would very fast blacklist their channels and
cause failed payments. As an initial step, we'd want the senders to
perform the non-strict decryption, eventually switching to the strict
decryption once adoption is sufficient.